pyproject.toml
[tool.poetry]
name = "cmem-plugin-yaml2json"
version = "0.0.0"
license = "Apache-2.0"
description = "Convert YAML input to JSON files"
authors = ["Natanael Arndt <natanael.arndt@eccenca.com>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
readme = "README-public.md"
keywords = [
    "eccenca Corporate Memory", "plugin"
]


[tool.poetry.dependencies]
# if you need to change python version here, change it also in .python-version
python = "^3.11"
ruamel-yaml = "^0.18.4"

[tool.poetry.dependencies.cmem-plugin-base]
version = "^4.1.0"
# allow-prereleases = true

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
coverage = "^7.2.3"
defusedxml = "^0.7.1"
flake8-formatter-junit-xml = "^0.0.6"
# https://github.com/smarie/python-genbadge/issues/31
pillow = "^9.5.0"
genbadge = "^1.1.0"
mypy = "^1.2.0"
# https://github.com/rasjani/pylint-junit/issues/1
pylint = "^2"
pylint-junit = "^0.3.2"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-memray = { version = "^1.4.0",  markers = "platform_system != 'Windows'" }
safety = "^1.10.3"
typed-ast = "^1.5.4"
wheel = "^0.38.4"
pip = ">=23.3"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
dirty = true

[tool.pylint.MASTER]
load-plugins="pylint_junit"

[tool.pylint.General]
ignore = "version.py"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
disable = "fixme"

# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#pylint
[tool.pylint.messages_control]
disable = "R0903"

[tool.pylint.format]
max-line-length = "88"

[tool.mypy]
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = ""

